Revert "Add "Find chat ID" panel to the ACP"
This was a huge commit that mixed a lot of refactoring and cleanup
with adding a sizable new feature. It also introduced a
migration (related to the new feature) that was broken and needs to be
completely redone.

After this revert, I will re-commit and fix that stuff in a cleaner
series of commits.

This reverts commit dc680542c0dc2b9478bd697741105ae502a9d466.